Improved Installer Script for Dockerfiles
The previous installer script had a few issues - it ran apt-get update every
time it was run, even if it had already been run earlier in the process. This
caused unnecessary delays in the build process. In addition, the script only
supported bash scripts, but not perl scripts.

To address these issues, the new installer script has been updated to only run
apt-get update once at the very beginning. It also now supports both types of
script - bash and perl - so that any type of installer can be run.

Additionally, a big title has been added to the build log to clearly show which
installer is being run at the moment, making it easier to track progress and
troubleshoot any issues that may arise.